fix: referencing a LICENSE or README outside the package directory now does not fail on install if the source file does not exist, but it will print a warning

+++ src/shut/renderers/setuptools.py
@@ -357,6 +357,9 @@ class SetuptoolsRenderer(Renderer[PackageModel]):
       def _tempcopy(src, dst):
         import atexit, shutil
         if not os.path.isfile(dst):
+          if not os.path.isfile(src):
+            print('warning: source file "{}" for destination "{}" does not exist'.format(src, dst))
+            return
           shutil.copyfile(src, dst)
           atexit.register(lambda: os.remove(dst))
     ''').lstrip())
